Trying to fix this lol

+++ nachos-3.4/code/userprog/addrspace.h
@@ -32,7 +32,7 @@ class AddrSpace {
 
     void SaveState();			// Save/restore address space-specific
     void RestoreState();		// info on a context switch
-    void copyIntoMemory(long, int);
+    void copyIntoMemory(int, int);
 
 
   private:
